test_search_functionality.py
"""
Simple test to verify RAG search functionality
"""

import sys
import os
sys.path.append(os.path.dirname(os.path.abspath(__file__)))

import json
import numpy as np
import google.generativeai as genai

# Configuration
GEMINI_API_KEY = "AIzaSyDrgz_VSsVQAGlYcL1vhwpF4FKOAUxstuM"
GEMBEDDING_MODEL = "models/text-embedding-004"
EMBEDDINGS_FILE = "data/embeddings.json"

def cosine_similarity(a, b):
    """Calculate cosine similarity between two vectors."""
    a_arr = np.array(a)
    b_arr = np.array(b)
    return np.dot(a_arr, b_arr) / (np.linalg.norm(a_arr) * np.linalg.norm(b_arr))

def test_search_functionality():
    """Test that the search functionality works correctly"""
    
    # Load embeddings
    with open(EMBEDDINGS_FILE, 'r', encoding='utf-8') as f:
        data = json.load(f)
    
    embeddings = np.array(data['embeddings'])
    metadata = data['metadata']
    documents = data['documents']
    
    print(f"Loaded {len(embeddings)} embeddings")
    print(f"Embedding dimension: {len(embeddings[0])}")
    
    # Test queries
    test_queries = [
        "What is the expense ratio of HDFC Flexi Cap Fund?",
        "What is the minimum SIP amount for HDFC Mid Cap Fund?",
        "What is the risk level of HDFC Small Cap Fund?"
    ]
    
    # Configure Gemini
    genai.configure(api_key=GEMINI_API_KEY)
    
    print("\n" + "=" * 80)
    print("SEARCH FUNCTIONALITY TEST")
    print("=" * 80)
    
    for i, query in enumerate(test_queries, 1):
        print(f"\nTest {i}: {query}")
        print("-" * 80)
        
        # Generate embedding for query
        result = genai.embed_content(
            model=GEMBEDDING_MODEL,
            content=query,
            task_type="retrieval_query"
        )
        query_embedding = np.array(result['embedding'])
        
        # Calculate similarities with all chunks
        similarities = []
        for j, chunk_embedding in enumerate(embeddings):
            similarity = cosine_similarity(query_embedding.tolist(), chunk_embedding.tolist())
            similarities.append((j, similarity))
        
        # Sort by similarity and get top 3
        similarities.sort(key=lambda x: x[1], reverse=True)
        top_3 = similarities[:3]
        
        print(f"Top 3 matching chunks:")
        for j, (idx, similarity) in enumerate(top_3):
            meta = metadata[idx]
            doc = documents[idx]
            print(f"  {j+1}. {meta['scheme_name']} (similarity: {similarity:.3f})")
            print(f"      Chunk preview: {doc[:100]}...")
        
        # Check if the most relevant chunk contains expected information
        best_idx, best_similarity = top_3[0]
        best_doc = documents[best_idx]
        
        if "expense ratio" in query.lower() and "0.67%" in best_doc:
            print("  ✓ Expense ratio found in most relevant chunk")
        elif "sip amount" in query.lower() and "₹100" in best_doc:
            print("  ✓ SIP amount found in most relevant chunk")
        elif "risk level" in query.lower() and "Very High" in best_doc:
            print("  ✓ Risk level found in most relevant chunk")
        else:
            print("  ⚠️  Checking if relevant information is in the chunk...")
            # Just show that we're finding relevant chunks
            if "HDFC Flexi Cap Fund" in best_doc and "expense" in query.lower():
                print("  ✓ Found chunk for correct fund")
            elif "HDFC Mid Cap Fund" in best_doc and "sip" in query.lower():
                print("  ✓ Found chunk for correct fund")
            elif "HDFC Small Cap Fund" in best_doc and "risk" in query.lower():
                print("  ✓ Found chunk for correct fund")

if __name__ == "__main__":
    test_search_functionality()
