requirements.txt
beautifulsoup4>=4.12.0
requests>=2.31.0
numpy>=1.24.0
google-generativeai>=0.8.0
fastapi>=0.111.0
uvicorn>=0.30.0
streamlit>=1.38.0